Fix benchmarks (#21)

+++ benchmarks/query/benchmark_unrepeated_two_component_query.py
@@ -20,11 +20,11 @@ import pytest
 
 
 class One(ecs.Component):
-    x: ecs.Float64
+    x: ecs.Float32
 
 
 class Two(ecs.Component):
-    x: ecs.Float64
+    x: ecs.Float32
 
 
 @pytest.mark.benchmark(
